@@ -1,23 +1,19 @@
-﻿using System.Collections.Generic;
-using Xamarin.Forms;
+﻿namespace SkiaSharp.Extended.UI.Controls;
 
-namespace SkiaSharp.Extended.UI.Controls
+[TypeConverter(typeof(Converters.SKConfettiColorCollectionTypeConverter))]
+public class SKConfettiColorCollection : List<Color>
 {
-	[TypeConverter(typeof(Converters.SKConfettiColorCollectionTypeConverter))]
-	public class SKConfettiColorCollection : List<Color>
+	public SKConfettiColorCollection()
 	{
-		public SKConfettiColorCollection()
-		{
-		}
+	}
 
-		public SKConfettiColorCollection(IEnumerable<Color> collection)
-			: base(collection)
-		{
-		}
+	public SKConfettiColorCollection(IEnumerable<Color> collection)
+		: base(collection)
+	{
+	}
 
-		public SKConfettiColorCollection(int capacity)
-			: base(capacity)
-		{
-		}
+	public SKConfettiColorCollection(int capacity)
+		: base(capacity)
+	{
 	}
 }
