@@ -1,18 +1,15 @@
-﻿using System.ComponentModel;
+﻿namespace SkiaSharp.Extended.UI.Controls;
 
-namespace SkiaSharp.Extended.UI.Controls
+[TypeConverter(typeof(Converters.SKConfettiPhysicsTypeConverter))]
+public readonly struct SKConfettiPhysics
 {
-	[TypeConverter(typeof(Converters.SKConfettiPhysicsTypeConverter))]
-	public readonly struct SKConfettiPhysics
+	public SKConfettiPhysics(double size, double mass)
 	{
-		public SKConfettiPhysics(double size, double mass)
-		{
-			Size = size;
-			Mass = mass;
-		}
-
-		public double Size { get; }
-
-		public double Mass { get; }
+		Size = size;
+		Mass = mass;
 	}
+
+	public double Size { get; }
+
+	public double Mass { get; }
 }
