@@ -1,21 +1,18 @@
-﻿using System.Collections.Generic;
+﻿namespace SkiaSharp.Extended.UI.Controls;
 
-namespace SkiaSharp.Extended.UI.Controls
+public class SKConfettiPhysicsCollection : List<SKConfettiPhysics>
 {
-	public class SKConfettiPhysicsCollection : List<SKConfettiPhysics>
+	public SKConfettiPhysicsCollection()
 	{
-		public SKConfettiPhysicsCollection()
-		{
-		}
+	}
 
-		public SKConfettiPhysicsCollection(IEnumerable<SKConfettiPhysics> collection)
-			: base(collection)
-		{
-		}
+	public SKConfettiPhysicsCollection(IEnumerable<SKConfettiPhysics> collection)
+		: base(collection)
+	{
+	}
 
-		public SKConfettiPhysicsCollection(int capacity)
-			: base(capacity)
-		{
-		}
+	public SKConfettiPhysicsCollection(int capacity)
+		: base(capacity)
+	{
 	}
 }
