@@ -1,21 +1,19 @@
-﻿using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
+﻿namespace SkiaSharp.Extended.UI.Controls.Converters;
 
-namespace SkiaSharp.Extended.UI.Controls.Converters
+#if XAMARIN_FORMS
+[TypeConversion(typeof(SKConfettiPhysics))]
+#endif
+public class SKConfettiPhysicsTypeConverter : ExtendedTypeConverter
 {
-	[TypeConversion(typeof(SKConfettiPhysics))]
-	public class SKConfettiPhysicsTypeConverter : TypeConverter
+	protected override object? Convert(string? value)
 	{
-		public override object? ConvertFromInvariantString(string? value)
-		{
-			if (value == null)
-				return null;
+		if (value == null)
+			return null;
 
-			value = value?.Trim();
+		value = value?.Trim();
 
-			var pointConverter = new PointTypeConverter();
-			var point = (Point)pointConverter.ConvertFromInvariantString(value);
-			return new SKConfettiPhysics(point.X, point.Y);
-		}
+		var pointConverter = new PointTypeConverter();
+		var point = (Point)pointConverter.ConvertFromInvariantString(value);
+		return new SKConfettiPhysics(point.X, point.Y);
 	}
 }
