@@ -1,21 +1,18 @@
-﻿using System.Collections.Generic;
+﻿namespace SkiaSharp.Extended.UI.Controls;
 
-namespace SkiaSharp.Extended.UI.Controls
+public class SKConfettiShapeCollection : List<SKConfettiShape>
 {
-	public class SKConfettiShapeCollection : List<SKConfettiShape>
+	public SKConfettiShapeCollection()
 	{
-		public SKConfettiShapeCollection()
-		{
-		}
+	}
 
-		public SKConfettiShapeCollection(IEnumerable<SKConfettiShape> collection)
-			: base(collection)
-		{
-		}
+	public SKConfettiShapeCollection(IEnumerable<SKConfettiShape> collection)
+		: base(collection)
+	{
+	}
 
-		public SKConfettiShapeCollection(int capacity)
-			: base(capacity)
-		{
-		}
+	public SKConfettiShapeCollection(int capacity)
+		: base(capacity)
+	{
 	}
 }
