@@ -1,38 +1,35 @@
-﻿using Xamarin.Forms;
+﻿namespace SkiaSharp.Extended.UI.Controls.Themes;
 
-namespace SkiaSharp.Extended.UI.Controls.Themes
+public partial class SKConfettiViewResources : ResourceDictionary
 {
-	public partial class SKConfettiViewResources : ResourceDictionary
+	private static bool registered;
+
+	public SKConfettiViewResources()
 	{
-		private static bool registered;
+		InitializeComponent();
+	}
 
-		public SKConfettiViewResources()
+	internal static void EnsureRegistered()
+	{
+		if (registered)
+			return;
+
+		var merged = Application.Current?.Resources?.MergedDictionaries;
+		if (merged != null)
 		{
-			InitializeComponent();
-		}
-
-		internal static void EnsureRegistered()
-		{
-			if (registered)
-				return;
-
-			var merged = Application.Current?.Resources?.MergedDictionaries;
-			if (merged != null)
+			foreach (var dic in merged)
 			{
-				foreach (var dic in merged)
+				if (dic.GetType() == typeof(SKConfettiViewResources))
 				{
-					if (dic.GetType() == typeof(SKConfettiViewResources))
-					{
-						registered = true;
-						break;
-					}
-				}
-
-				if (!registered)
-				{
-					merged.Add(new SKConfettiViewResources());
 					registered = true;
+					break;
 				}
+			}
+
+			if (!registered)
+			{
+				merged.Add(new SKConfettiViewResources());
+				registered = true;
 			}
 		}
 	}
